pdfchat/requirements.txt
langchain
langchain_openai
langchain_community
langchain_text_splitters
langchain_core
langchain_experimental
faiss-cpu
pymupdf
pypdf
flask
python-dotenv
flask-cors
pandas
tabulate
pandasai
PyPDF 
PyPDF2 
python-docx
langchain_google_genai
google-generativeai
crewai